Split service rule tests by theme
